+++ lib/redis.rb
@@ -764,7 +764,7 @@ class Redis
   # Set one or more values, only if none of the keys exist.
   #
   # @example
-  #   redis.msetnx({ "key1" => "v1", "key2" => "v2" })
+  #   redis.mapped_msetnx({ "key1" => "v1", "key2" => "v2" })
   #     # => true
   #
   # @param [Hash] hash keys mapping to values
@@ -1620,6 +1620,28 @@ class Redis
     end
   end
 
+  # Return a range of members with the same score in a sorted set, by reversed lexicographical ordering.
+  # Apart from the reversed ordering, #zrevrangebylex is similar to #zrangebylex.
+  #
+  # @example Retrieve members matching a
+  #   redis.zrevrangebylex("zset", "[a", "[a\xff")
+  #     # => ["abbygail", "abby", "abagael", "aaren"]
+  # @example Retrieve the last 2 members matching a
+  #   redis.zrevrangebylex("zset", "[a", "[a\xff", :limit => [0, 2])
+  #     # => ["abbygail", "abby"]
+  #
+  # @see #zrangebylex
+  def zrevrangebylex(key, max, min, options = {})
+    args = []
+
+    limit = options[:limit]
+    args.concat(["LIMIT"] + limit) if limit
+
+    synchronize do |client|
+      client.call([:zrevrangebylex, key, max, min] + args)
+    end
+  end
+
   # Return a range of members in a sorted set, by score.
   #
   # @example Retrieve members with score `>= 5` and `< 100`
@@ -1895,7 +1917,7 @@ class Redis
   # Get the values of all the given hash fields.
   #
   # @example
-  #   redis.hmget("hash", "f1", "f2")
+  #   redis.mapped_hmget("hash", "f1", "f2")
   #     # => { "f1" => "v1", "f2" => "v2" }
   #
   # @param [String] key
@@ -2032,7 +2054,7 @@ class Redis
     end
   end
 
-  # Inspect the state of the Pub/Sub subsystem. 
+  # Inspect the state of the Pub/Sub subsystem.
   # Possible subcommands: channels, numsub, numpat.
   def pubsub(subcommand, *args)
     synchronize do |client|
@@ -2458,7 +2480,7 @@ class Redis
   # Scan a set
   #
   # @example Retrieve all of the keys in a set
-  #   redis.sscan("set").to_a
+  #   redis.sscan_each("set").to_a
   #   # => ["key1", "key2", "key3"]
   #
   # @param [Hash] options