+++ lib/redis/connection/hiredis.rb
@@ -9,11 +9,12 @@ class Redis
 
       def self.connect(config)
         connection = ::Hiredis::Connection.new
+        connect_timeout = (config.fetch(:connect_timeout, 0) * 1_000_000).to_i
 
         if config[:scheme] == "unix"
-          connection.connect_unix(config[:path], Integer(config[:connect_timeout] * 1_000_000))
+          connection.connect_unix(config[:path], connect_timeout)
         else
-          connection.connect(config[:host], config[:port], Integer(config[:connect_timeout] * 1_000_000))
+          connection.connect(config[:host], config[:port], connect_timeout)
         end
 
         instance = new(connection)