+++ CHANGELOG.md
@@ -20,6 +20,11 @@
 
 * Added `:connect_timeout` option.
 
+* Added support for `:limit` option for `ZREVRANGEBYLEX`.
+
+* Fixed an issue where connections become inconsistent when using Ruby's
+  Timeout module outside of the client (see #501, #502).
+
 # 3.2.0
 
 * Redis Sentinel support.